+++ kit/dapp/src/lib/queries/user/user-fragment.ts
@@ -6,7 +6,6 @@ export const AccountFragment = theGraphGraphqlKit(`
     id
     lastActivity
     balancesCount
-    activityEventsCount
   }
 `);
