+++ kit/subgraph/src/assets/calculations/fixed-yield.ts
@@ -1,5 +1,5 @@
 import { Address, BigInt, log } from "@graphprotocol/graph-ts";
-import { Bond } from "../../../generated/schema";
+import { Bond, FixedYield } from "../../../generated/schema";
 import { FixedYield as FixedYieldContract } from "../../../generated/templates/FixedYield/FixedYield";
 import { toDecimals } from "../../utils/decimals";
 import { fetchFixedYield, fetchFixedYieldPeriod } from "../fetch/fixed-yield";
@@ -43,4 +43,35 @@ export function updateFixedYield(bond: Bond): void {
     fixedYieldPeriod.totalYield.toString(),
     fixedYieldPeriod.totalYieldExact.toString(),
   ]);
+}
+
+/**
+ * Updates the underlying balance of a FixedYield contract
+ *
+ * @param address - Address of the FixedYield contract
+ * @param amount - Amount to add to the underlying balance
+ * @param decimals - Decimals of the underlying asset
+ * @returns The updated FixedYield entity or null if not found
+ */
+export function updateFixedYieldBalance(
+  address: Address,
+  amount: BigInt,
+  decimals: i32
+): FixedYield | null {
+  const fixedYield = FixedYield.load(address);
+
+  if (!fixedYield) {
+    return null;
+  }
+
+  // Update the balances
+  fixedYield.underlyingBalanceExact =
+    fixedYield.underlyingBalanceExact.plus(amount);
+  fixedYield.underlyingBalance = toDecimals(
+    fixedYield.underlyingBalanceExact,
+    decimals
+  );
+  fixedYield.save();
+
+  return fixedYield;
 }
