+++ kit/subgraph/src/assets/fetch/cryptocurrency.ts
@@ -1,7 +1,7 @@
 import { Address, BigDecimal, BigInt } from "@graphprotocol/graph-ts";
 import { CryptoCurrency } from "../../../generated/schema";
 import { CryptoCurrency as CryptoCurrencyContract } from "../../../generated/templates/CryptoCurrency/CryptoCurrency";
-import { fetchAccount } from "../../fetch/account";
+import { fetchAccount } from "../../utils/account";
 import { AssetType } from "../../utils/enums";
 
 export function fetchCryptoCurrency(address: Address): CryptoCurrency {
