+++ kit/subgraph/src/assets/fetch/equity.ts
@@ -1,7 +1,7 @@
 import { Address, BigDecimal, BigInt } from "@graphprotocol/graph-ts";
 import { Equity } from "../../../generated/schema";
 import { Equity as EquityContract } from "../../../generated/templates/Equity/Equity";
-import { fetchAccount } from "../../fetch/account";
+import { fetchAccount } from "../../utils/account";
 import { AssetType } from "../../utils/enums";
 
 export function fetchEquity(address: Address): Equity {
