+++ kit/subgraph/src/assets/fetch/stablecoin.ts
@@ -1,7 +1,7 @@
 import { Address, BigDecimal, BigInt } from "@graphprotocol/graph-ts";
 import { StableCoin } from "../../../generated/schema";
 import { StableCoin as StableCoinContract } from "../../../generated/templates/StableCoin/StableCoin";
-import { fetchAccount } from "../../fetch/account";
+import { fetchAccount } from "../../utils/account";
 import { AssetType } from "../../utils/enums";
 
 export function fetchStableCoin(address: Address): StableCoin {
