+++ kit/subgraph/src/assets/handlers/burn.ts
@@ -7,9 +7,9 @@ import {
   store,
 } from "@graphprotocol/graph-ts";
 import { Account, AssetBalance } from "../../../generated/schema";
-import { fetchAccount } from "../../fetch/account";
-import { createActivityLogEntry, EventType } from "../../fetch/activity-log";
-import { fetchAssetBalance } from "../../fetch/balance";
+import { fetchAccount } from "../../utils/account";
+import { createActivityLogEntry, EventType } from "../../utils/activity-log";
+import { fetchAssetBalance } from "../../utils/balance";
 import { decrease, increase } from "../../utils/counters";
 import { setValueWithDecimals } from "../../utils/decimals";
 import { AssetType } from "../../utils/enums";
