+++ kit/subgraph/src/assets/stablecoin.ts
@@ -14,7 +14,7 @@ import {
   UserBlocked,
   UserUnblocked,
 } from "../../generated/templates/StableCoin/StableCoin";
-import { createActivityLogEntry, EventType } from "../fetch/activity-log";
+import { createActivityLogEntry, EventType } from "../utils/activity-log";
 import { AssetType } from "../utils/enums";
 import { calculateCollateral } from "./calculations/collateral";
 import { calculateConcentration } from "./calculations/concentration";
