+++ kit/subgraph/src/distribution/vesting-airdrop.ts
@@ -30,7 +30,7 @@ import {
   TokensWithdrawn,
 } from "../../generated/templates/VestingAirdropTemplate/VestingAirdrop";
 
-import { fetchAccount } from "../fetch/account";
+import { fetchAccount } from "../utils/account";
 import { toDecimals } from "../utils/decimals";
 
 // Helper function to get token decimals from any asset type
