+++ kit/subgraph/src/factories/equity-factory.ts
@@ -2,8 +2,8 @@ import { EquityCreated } from "../../generated/EquityFactory/EquityFactory";
 import { Equity } from "../../generated/templates";
 import { fetchAssetCount } from "../assets/fetch/asset-count";
 import { fetchEquity } from "../assets/fetch/equity";
-import { fetchAccount } from "../fetch/account";
-import { createActivityLogEntry, EventType } from "../fetch/activity-log";
+import { fetchAccount } from "../utils/account";
+import { createActivityLogEntry, EventType } from "../utils/activity-log";
 import { AssetType, FactoryType } from "../utils/enums";
 import { fetchFactory } from "./fetch/factory";
