+++ kit/subgraph/src/factories/fetch/dvpswap-factory.ts
@@ -1,6 +1,6 @@
 import { Address } from "@graphprotocol/graph-ts";
 import { DvPSwapFactory } from "../../../generated/schema";
-import { fetchAccount } from "../../fetch/account";
+import { fetchAccount } from "../../utils/account";
 import { FactoryType } from "../../utils/enums";
 export function fetchDvPSwapFactory(address: Address): DvPSwapFactory {
   let factory = DvPSwapFactory.load(address);
