+++ kit/subgraph/src/factories/stablecoin-factory.ts
@@ -2,8 +2,8 @@ import { StableCoinCreated } from "../../generated/StableCoinFactory/StableCoinF
 import { StableCoin } from "../../generated/templates";
 import { fetchAssetCount } from "../assets/fetch/asset-count";
 import { fetchStableCoin } from "../assets/fetch/stablecoin";
-import { fetchAccount } from "../fetch/account";
-import { createActivityLogEntry, EventType } from "../fetch/activity-log";
+import { fetchAccount } from "../utils/account";
+import { createActivityLogEntry, EventType } from "../utils/activity-log";
 import { AssetType, FactoryType } from "../utils/enums";
 import { fetchFactory } from "./fetch/factory";
