+++ kit/subgraph/src/factories/xvp-settlement-factory.ts
@@ -1,7 +1,7 @@
 import { XvPSettlement } from "../../generated/templates";
 import { XvPSettlementCreated } from "../../generated/XvPSettlementFactory/XvPSettlementFactory";
-import { fetchAccount } from "../fetch/account";
-import { fetchXvPSettlement } from "../fetch/xvp-settlement";
+import { fetchXvPSettlement } from "../trading/fetch/xvp-settlement";
+import { fetchAccount } from "../utils/account";
 
 /**
  * Handles XvPSettlementCreated events from the XvPSettlementFactory contract.
