+++ kit/subgraph/src/utils/account.ts
@@ -14,7 +14,6 @@ export function fetchAccount(address: Address): Account {
     account.pausedBalancesCount = 0;
     account.pausedBalanceExact = BigInt.zero();
     account.pausedBalance = toDecimals(account.pausedBalanceExact, 18);
-    account.activityEventsCount = 0;
     if (ethereum.hasCode(address).inner) {
       account.isContract = true;
     } else {
